cryptotrades/requirements.txt
# === Core Dependencies (required) ===
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
websocket-client>=1.6.0

# === Optional: Enhanced features ===
# torch — RL agent DQN (falls back to Q-table without it)
# coinbase — Coinbase spot trading API
# transformers — HuggingFace NLP sentiment (falls back to RSS)
# feedparser — RSS news feed parsing
